users-service/requirements.txt
fastapi>=0.97.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.16
pydantic[email]>=1.10.9
pydantic-settings
alembic>=1.11.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
mysqlclient
